+++ src/Lira.Domain/Handling/Generating/TextPartsProvider.cs
@@ -1,10 +1,12 @@
 namespace Lira.Domain.Handling.Generating;
 
-public class TextPartsProvider(IEnumerable<ITextParts> valueParts)
+public class TextPartsProvider(IEnumerable<ITextParts> parts)
 {
+    public readonly static  TextPartsProvider Empty = new([]);
+
     public IEnumerable<string> Get(RuleExecutingContext ctx)
     {
-        foreach (var part in valueParts)
+        foreach (var part in parts)
         {
             foreach (var text in part.Get(ctx))
             {