+++ src/Lira.Domain/HttpContextData.cs
@@ -1,3 +1,3 @@
 namespace Lira.Domain;
 
-public record HttpContextData(RuleExecutingContext RuleExecutingContext, ResponseData Response);
+record HttpContextData(RuleExecutingContext RuleExecutingContext, ResponseData Response);