+++ src/Lira/Middlewares/LoggingMiddleware.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text;
+using Lira.Domain.Utils;
 
 namespace Lira.Middlewares;
 
@@ -93,7 +94,8 @@ public class LoggingMiddleware
         response.Body.Seek(0, SeekOrigin.Begin);
 
         //...and copy it into a string
-        string text = await new StreamReader(response.Body).ReadToEndAsync();
+
+        string text = await new StreamReader(response.Body, response.TryGetContentEncoding() ?? Encoding.UTF8).ReadToEndAsync();
         text = text.Replace("\r\n", "\n").Replace("\n", Environment.NewLine);
 
         //We need to reset the reader for the response so that the client can read it.